requirements.txt
asgiref==3.8.1
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.7.0
django-filter==25.1
django-money==3.5.4
djangorestframework==3.15.2
djangorestframework-datatables==0.7.2
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
py-moneyed==3.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
RapidFuzz==3.12.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
thefuzz==0.22.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
